refactor before writing tests
