+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/view/AggregationConfiguration.java
@@ -7,7 +7,6 @@ package io.opentelemetry.sdk.metrics.view;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.api.metrics.Instrument;
-import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
 
 @AutoValue
@@ -19,11 +18,9 @@ public abstract class AggregationConfiguration {
   }
 
   /** Which {@link Aggregation} should be used for this View. */
-  @Nullable
   public abstract Aggregation aggregation();
 
   /** What {@link Temporality} should be used for this View (delta vs. cumulative). */
-  @Nullable
   public abstract Temporality temporality();
 
   /** An enumeration which describes the time period over which metrics should be aggregated. */
